validaciones

+++ src/main/java/com/example/service/DepartamentoServiceImpl.java
@@ -10,11 +10,14 @@ import io.quarkus.panache.common.Page;
 import io.quarkus.panache.common.Parameters;
 import jakarta.enterprise.context.ApplicationScoped;
 import jakarta.inject.Inject;
+import jakarta.json.Json;
 import jakarta.ws.rs.core.Response;
 
 import java.net.URI;
 import java.util.List;
+import java.util.Map;
 import java.util.NoSuchElementException;
+import java.util.Optional;
 
 @ApplicationScoped
 public class DepartamentoServiceImpl implements DepartamentoService {
@@ -59,12 +62,16 @@ public class DepartamentoServiceImpl implements DepartamentoService {
     public Response insert(DepartamentoDto dto) {
         var entity = new Departamento();
         mapper.toEntity(entity, dto); //Conversion de DTO -> Entity
-        departamentoRepository.persist(entity); //Insercion
-        //Al terminar el proceso se espera que devuelva el DTO del registro insertado
-        return Response.created(
-                        URI.create("/departamentos/" + entity.getId()))
-                .entity(mapper.toDTO(entity)) //Conversión del registro insertado a DTO
-                .build();
+       // if(departamentoRepository.find("codigo=?",dto.codigo()).count()>0){
+            departamentoRepository.persist(entity); //Insercion
+            //Al terminar el proceso se espera que devuelva el DTO del registro insertado
+            return Response.created(
+                            URI.create("/departamentos/" + entity.getId()))
+                    .entity(mapper.toDTO(entity)) //Conversión del registro insertado a DTO
+                    .build();
+       // }
+       // return Response.status(400).entity(Map.of("message", "Error desconocido.")).build();
+
     }
 
     @Override
@@ -72,16 +79,35 @@ public class DepartamentoServiceImpl implements DepartamentoService {
         var original = departamentoRepository.findByIdOptional(id)
                 .orElseThrow(() -> new NoSuchElementException("No se encontró registro"));
 
-        mapper.toEntity(original,dto);
+        mapper.toEntity(original, dto);
         departamentoRepository.persist(original);
         return Response.ok().entity(mapper.toDTO(original)).build();
     }
 
     @Override
     public Response delete(long id) {
-        if(departamentoRepository.deleteById(id)){
-            return Response.ok()/*.entity("Operación exitosa.")*/.build();
+        Optional<Departamento> departamentoOpt = Optional.ofNullable(departamentoRepository.findById(id));
+
+        if (departamentoOpt.isEmpty()) {
+            return Response.status(400)
+                    .entity(Map.of("message", "No se encontró registro."))
+                    .build();
         }
-        return Response.status(400)/*.entity("No se encontró registro")*/.build();
+
+        Departamento departamento = departamentoOpt.get();
+
+        if (departamento.getMunicipios() != null && !departamento.getMunicipios().isEmpty()) {
+            return Response.status(400)
+                    .entity(Map.of("message", "Este departamento no puede ser eliminado porque tiene municipios asociados."))
+                    .build();
+        }
+
+        if (departamentoRepository.deleteById(id)) {
+            return Response.ok(Map.of("message", "Operación exitosa: departamento eliminado.")).build();
+        }else{
+            return Response.status(400).entity(Map.of("message", "Error desconocido.")).build();
+
+        }
+
     }
 }
